json to csv. and update file structure
